+++ src/pages/main/forum/Forum.jsx
@@ -5,8 +5,7 @@ import RightSidebar from "./RightSidebar";
 import { Link, useLocation } from "react-router-dom";
 import { getAllForumPostAPI, getAllForumTopicsAPI, likeOrUnlikePostAPI, savePostAPI, unsavePostAPI } from "~/apis";
 import { toast } from "react-toastify";
-import ReactQuill from "react-quill-new";
-import "react-quill-new/dist/quill.snow.css";
+import CreatePostModal from "~/components/forum/CreatePostModal";
 
 const Forum = () => {
   const [posts, setPosts] = useState([]);
@@ -71,88 +70,6 @@ const Forum = () => {
     return new File([u8arr], filename, { type: mime });
   };
 
-  // Handle publish new post
-  const handlePublish = async () => {
-    let loadingToastId = null;
-    try {
-      setIsLoading(true);
-      loadingToastId = toast.loading("Posting your article...");
-
-      if (!title.trim() || !content.trim()) {
-        toast.update(loadingToastId, {
-          render: "Title and content cannot be empty!",
-          type: "error",
-          isLoading: false,
-          autoClose: 3000,
-        });
-        return;
-      }
-
-      if (!selectedTopicId || isNaN(selectedTopicId)) {
-        toast.update(loadingToastId, {
-          render: "You must select a valid topic!",
-          type: "error",
-          isLoading: false,
-          autoClose: 3000,
-        });
-        return;
-      }
-
-      const { div, imgTags, base64Images } = extractBase64ImagesFromHTML(content);
-
-      const allImages = [];
-      if (coverImageFile) allImages.push(coverImageFile);
-      const contentImagesAsFiles = base64Images.map((base64, i) =>
-        convertBase64ToFile(base64, `content_img_${i}.png`)
-      );
-      allImages.push(...contentImagesAsFiles);
-
-      let uploadedUrls = [];
-      if (allImages.length > 0) {
-        uploadedUrls = await uploadImageAPI(allImages);
-      }
-
-      imgTags.forEach((img) => img.remove());
-      const newContent = div.innerHTML;
-
-      const payload = {
-        title,
-        content: newContent,
-        images: uploadedUrls,
-        forumTopicTypeId: Number(selectedTopicId),
-      };
-
-      await postForumAPI(payload);
-
-      toast.update(loadingToastId, {
-        render: "Post published successfully!",
-        type: "success",
-        isLoading: false,
-        autoClose: 3000,
-      });
-
-      // Reset form and refresh posts
-      setTitle("");
-      setContent("");
-      setCoverImagePreview(null);
-      setCoverImageFile(null);
-      setSelectedTopicId(null);
-      setExpandForm(false); // Auto close modal after create
-      // Refresh posts
-      const res = await getAllForumPostAPI();
-      setPosts(res.data);
-    } catch (err) {
-      console.error(err);
-      toast.update(loadingToastId, {
-        render: "Failed to publish post. Please try again.",
-        type: "error",
-        isLoading: false,
-        autoClose: 3000,
-      });
-    } finally {
-      setIsLoading(false);
-    }
-  };
 
   useEffect(() => {
     const fetchPosts = async () => {
@@ -336,134 +253,14 @@ const Forum = () => {
           </div>
 
           {/* Modal for New Post Form */}
-          {expandForm && (
-            <div
-              className="fixed inset-0 z-30 flex items-start justify-center pt-16"
-              style={{
-                backgroundColor: 'rgba(50, 50, 50, 0.7)',
-                backdropFilter: 'blur(4px)',
-                WebkitBackdropFilter: 'blur(4px)'
-              }}
-            >
-              <div className="bg-white dark:bg-[#1e1e2f] rounded-xl shadow-lg p-8 w-full max-w-2xl max-h-[calc(100vh-8rem)] overflow-y-auto relative mx-4 mt-10">
-                <button
-                  className="absolute top-3 right-3 text-gray-500 hover:text-gray-800 dark:text-gray-400 dark:hover:text-white"
-                  onClick={() => setExpandForm(false)}
-                  title="Close"
-                >
-                  <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" strokeWidth={1.5} stroke="currentColor" className="w-6 h-6">
-                    <path strokeLinecap="round" strokeLinejoin="round" d="M6 18L18 6M6 6l12 12" />
-                  </svg>
-                </button>
-
-                <h2 className="text-2xl font-bold mb-6 text-gray-800 dark:text-white">✍️ Create New Post</h2>
-
-                {/* Cover Image */}
-                <div className="mb-6">
-                  <label className="block mb-2 text-sm font-medium text-gray-700 dark:text-neutral-300">Cover Image</label>
-                  <input
-                    type="file"
-                    accept="image/*"
-                    onChange={handleImageUpload}
-                    className="file:mr-4 file:py-2 file:px-4 file:rounded-md file:border-0 file:text-sm file:bg-blue-600 file:text-white hover:file:bg-blue-700 cursor-pointer dark:file:bg-blue-500 dark:hover:file:bg-blue-600"
-                  />
-                  {coverImagePreview && (
-                    <img
-                      src={coverImagePreview}
-                      alt="Preview"
-                      className="mt-4 rounded-lg max-h-64 object-cover w-full border border-gray-300 dark:border-neutral-700"
-                    />
-                  )}
-                </div>
-
-                {/* Topic Select */}
-                <div className="mb-6">
-                  <label className="block mb-2 text-sm font-medium text-gray-700 dark:text-neutral-300">Select Topic</label>
-                  <select
-                    value={selectedTopicId ?? ""}
-                    onChange={(e) => setSelectedTopicId(Number(e.target.value))}
-                    className="w-full p-3 rounded-md border border-gray-300 dark:border-neutral-700 bg-white dark:bg-[#181818] text-gray-700 dark:text-neutral-200 text-sm focus:ring-2 focus:ring-blue-500 outline-none transition-all"
-                  >
-                    <option value="" disabled>--Select Topic--</option>
-                    {topics.length === 0 ? (
-                      <option disabled>No topics available</option>
-                    ) : (
-                      topics.map((topic) => (
-                        <option key={topic.id} value={Number(topic.id)}>
-                          {topic.title}
-                        </option>
-                      ))
-                    )}
-                  </select>
-                </div>
-
-                {/* Title */}
-                <input
-                  type="text"
-                  placeholder="📝 Post title..."
-                  value={title}
-                  onChange={(e) => setTitle(e.target.value)}
-                  className="w-full p-4 mb-5 text-xl font-semibold rounded-md border border-gray-300 dark:border-neutral-700 bg-white dark:bg-[#181818] text-gray-800 dark:text-neutral-100 focus:ring-2 focus:ring-blue-500 outline-none transition-all"
-                />
-
-                {/* Content Editor */}
-                <ReactQuill
-                  theme="snow"
-                  value={content}
-                  onChange={setContent}
-                  placeholder="✍️ Write your post content here..."
-                  className="mb-6 bg-white dark:bg-[#181818] text-black dark:text-white border-none rounded-md"
-                  modules={{
-                    toolbar: [
-                      [{ header: [1, 2, 3, false] }],
-                      ["bold", "italic", "underline"],
-                      ["blockquote", "code-block"],
-                      [{ list: "ordered" }, { list: "bullet" }],
-                      ["link", "image"],
-                      ["clean"],
-                    ],
-                  }}
-                />
-
-                {/* Publish Button */}
-                <div className="flex justify-between items-center mt-10">
-                  <button
-                    onClick={handlePublish}
-                    disabled={isLoading}
-                    className={`flex items-center justify-center gap-2 bg-blue-600 dark:bg-blue-500 hover:bg-blue-700 dark:hover:bg-blue-600 text-white font-semibold px-6 py-2 rounded-lg transition-all duration-300 shadow-md hover:shadow-lg ${isLoading ? "opacity-50 cursor-not-allowed" : ""}`}
-                  >
-                    {isLoading ? (
-                      <>
-                        <svg
-                          className="animate-spin h-5 w-5 text-white"
-                          xmlns="http://www.w3.org/2000/svg"
-                          fill="none"
-                          viewBox="0 0 24 24"
-                        >
-                          <circle
-                            className="opacity-25"
-                            cx="12"
-                            cy="12"
-                            r="10"
-                            stroke="currentColor"
-                            strokeWidth="4"
-                          ></circle>
-                          <path
-                            className="opacity-75"
-                            fill="currentColor"
-                            d="M4 12a8 8 0 018-8v4a4 4 0 00-4 4H4z"
-                          ></path>
-                        </svg>
-                        <span>Posting...</span>
-                      </>
-                    ) : (
-                      <span>🚀 Publish</span>
-                    )}
-                  </button>
-                </div>
-              </div>
-            </div>
-          )}
+          <CreatePostModal
+            isOpen={expandForm}
+            onClose={() => setExpandForm(false)}
+            onPostCreated={async () => {
+              const res = await getAllForumPostAPI();
+              setPosts(res.data);
+            }}
+          />
 
           {/* Forum Posts List */}
           {posts.length === 0 ? (
